Finished User Component
https://trello.com/c/D3uMPuu2/2-create-user-component

I finished the User component. This entailed creating a styled list of social media links via flexbox, styling a 'Hire Me' button, positioning the elements relative to each other on the page, adjusting margin values to ensure a symmetrical UI, and adjusting the grid so that the site is sexy regardless of screen size.

* added better_errors gem
* added btn-user class to style 'Hire Me' button
* added ul for social media icons
* organized icons by professional vs. fun profiles
* placed all content within user-content div so that entire component adjusts w/ screen size changes
* adjusted padding/margin values for user-content, priority-icons, secondary-icons
* removed default padding on row so content fits page perfectly
* created nested properties within user class to style ul, li, and p