Finished Projects Component
https://trello.com/c/9J0dNQU8/3-create-projects-component

I finished the projects component by adding a scrollable div that contains cards corresponding to each of my projects. Each card contains links to the site, the Github repo, and the Trello board used to manage my work. Furthermore, the site is entirely responsive, and the content is wrapped in a Materialize grid.

* added a 'content-base' class to contain all of the cards.
* used materialize's default card structure, then overrode necessary properties by creating associated override classes, i.e. 'override-card-actions' overrides the 'card-action' class.
* updated README with instructions on how to run project in development
* added basic jquery to adjust z-index of projects container
* will need to revisit component during polishing in order to add content and imgs