Finished Main Work on Navbar
https://trello.com/c/ZkjUwGCv/7-create-navbar-component

While getting ready to build the CV section, I realized there wasn't a good way navigate through the site, so I dropped in Materialize's navbar and fixed it within the main content div. Upon clicking a button on the main-page, the app will render the appropriate ERB partial complete with a fixed navbar. I aligned the icons, ensured the navbar was responsive, and made it play nicely with existing content.

I also refactored the jQuery so that it renders ERB partials based on btn clicks. Unfortunately, the jQuery no longer lives in individual js files, and has to be included in the view template for it to work. Also, any jQuery that lives on the navbar has to be included in the partial that actually contains the code for the navbar.

* created override classes for the navbar in _navbar.scss
* adjusted size of .content-base so that it takes up the entire page width
* removed js files for btn jquery
* moved btn jquery to index.html.erb
* partials are now being rendered according to btn click
* moved projects navbar jquery to projects partial
* moved cv navbar jquery to cv partial